+++ crates/sqlite-store/src/note/tests.rs
@@ -105,7 +105,7 @@ fn create_consumed_input_note_with_consumer(
 async fn insert_input_notes_with_tx_order(
     store: &crate::SqliteStore,
     notes: &[InputNoteRecord],
-    consumed_tx_order: Option<u32>,
+    consumed_tx_order: Option<u16>,
 ) {
     let notes = notes.to_vec();
     store
